Working program but needs some serious refactoring.

+++ ImageKml/Models/GeoData.cs
@@ -10,8 +10,8 @@ namespace ImageKml.Models
     {
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public int Altitude { get; set; }
-        public int LatitudeSpan { get; set; }
-        public int LongitudeSpan { get; set; }
+        public float Altitude { get; set; }
+        public float LatitudeSpan { get; set; }
+        public float LongitudeSpan { get; set; }
     }
 }
